hci_le_set_ext_scan extracted
